Move `thenablify` to same file
